Groups

+++ core/mutation.py
@@ -2,8 +2,14 @@ import graphene
 import reversion
 from django.contrib.contenttypes.models import ContentType
 from core.lib import get_id
-from .models import Comment, Group
+from .models import Comment, Group, User
 from .nodes import Node, GroupNode, CommentNode
+from django.db import transaction
+from .exceptions import *
+import logging
+
+# Get an instance of a logger
+logger = logging.getLogger(__name__)
 
 class CommentInput(graphene.InputObjectType):
     description = graphene.String(required=True)
@@ -93,6 +99,7 @@ class GroupInput(graphene.InputObjectType):
     name = graphene.String(required=True)
     description = graphene.String(required=True)
     is_open = graphene.Boolean(required=True)
+    is_2fa_required = graphene.Boolean(required=True)
     tags = graphene.List(graphene.NonNull(graphene.String))
 
 class CreateGroup(graphene.Mutation):
@@ -104,19 +111,27 @@ class CreateGroup(graphene.Mutation):
 
     def mutate(self, info, input):
         try:
+            if not info.context.user.is_authenticated:
+                raise UserNotLoggedIn
+
             with reversion.create_revision():
                 group = Group.objects.create(
                     name=input['name'],
                     description=input['description'],
                     is_open=input['is_open'],
+                    is_2fa_required=input['is_2fa_required'],
                     tags=input['tags'],
                 )
+
+                #add creator as group owner
+                group.join(info.context.user, 'owner')
 
                 reversion.set_user(info.context.user)
                 reversion.set_comment("createGroup mutation")
 
             ok = True
-        except:
+        except Exception as e:
+            logger.error('Exception in CreateGroup: {}.'.format(e))
             group = None
             ok = False
 
@@ -132,11 +147,16 @@ class UpdateGroup(graphene.Mutation):
 
     def mutate(self, info, id, input):
         try:
+            group = Group.objects.get(pk=get_id(id))
+
+            if not group.can_change(info.context.user):
+                raise UserNotAuthorized()
+
             with reversion.create_revision():
-                group = Group.objects.get(pk=get_id(id))
                 group.name = input['name']
                 group.description = input['description']
                 group.is_open=input['is_open']
+                group.is_2fa_required=input['is_2fa_required']
                 group.tags=input['tags']
                 group.save()
 
@@ -144,7 +164,8 @@ class UpdateGroup(graphene.Mutation):
                 reversion.set_comment("updateGroup mutation")
 
             ok = True
-        except:
+        except Exception as e:
+            logger.error('Exception in UpdateGroup: {}.'.format(e))
             group = None
             ok = False
 
@@ -158,15 +179,27 @@ class DeleteGroup(graphene.Mutation):
 
     def mutate(self, info, id):
         try:
+            group = Group.objects.get(pk=get_id(id))
+
+            if not group.can_change(info.context.user):
+                raise UserNotAuthorized()
+
             with reversion.create_revision():
-                group = Group.objects.get(pk=get_id(id))
-                group.delete()
+                if group.members.exclude(user=info.context.user).exists():
+                    #other members exist, cannot delete group
+                    raise GroupContainsMembers
+                with transaction.atomic():
+                    #owner will only leave group to clear all members from group.
+                    #if group.delete fails, owner membership delete will be rollbacked
+                    group.leave(info.context.user)
+                    group.delete()
 
                 reversion.set_user(info.context.user)
                 reversion.set_comment("deleteGroup mutation")
 
             ok = True
-        except:
+        except Exception as e:
+            logger.error('Exception in DeleteGroup: {}.'.format(e))
             ok = False
 
         return DeleteGroup(ok=ok)
@@ -179,11 +212,24 @@ class JoinGroup(graphene.Mutation):
     group = graphene.Field(lambda: GroupNode)
 
     def mutate(self, info, id):
-        group = Group.objects.get(pk=get_id(id))
-        if group.can_join(info.context.user):
-            group.join(info.context.user)
+        try:
+            group = Group.objects.get(pk=get_id(id))
+
+            if not group.can_join(info.context.user):
+                raise UserNotAuthorized
+
+            with reversion.create_revision():
+                if group.is_open:
+                    group.join(info.context.user, 'member')
+                else:
+                    group.join(info.context.user, 'pending')
+
+                reversion.set_user(info.context.user)
+                reversion.set_comment("joinGroup mutation")
+
             ok = True
-        else:
+        except Exception as e:
+            logger.error('Exception in JoinGroup: {}.'.format(e))
             ok = False
 
         return JoinGroup(ok=ok, group=group)
@@ -196,15 +242,93 @@ class LeaveGroup(graphene.Mutation):
     group = graphene.Field(lambda: GroupNode)
 
     def mutate(self, info, id):
-        group = Group.objects.get(pk=get_id(id))
+        try:
+            group = Group.objects.get(pk=get_id(id))
 
-        if info.context.user.is_authenticated:
-            group.leave(info.context.user)
+            if not info.context.user.is_authenticated:
+                raise UserNotLoggedIn
+
+            with reversion.create_revision():
+                group.leave(info.context.user)
+
+                reversion.set_user(info.context.user)
+                reversion.set_comment("leaveGroup mutation")
+
             ok = True
-        else:
+        except Exception as e:
+            logger.error('Exception in LeaveGroup: {}.'.format(e))
             ok = False
 
         return LeaveGroup(ok=ok, group=group)
+
+class MembershipInput(graphene.InputObjectType):
+    userid = graphene.ID(required=True)
+    type = graphene.String(required=True)
+
+class ChangeMembershipGroup(graphene.Mutation):
+
+    class Arguments:
+        id = graphene.ID(required=True)
+        input = MembershipInput(required=True)
+
+    ok = graphene.Boolean()
+    group = graphene.Field(lambda: GroupNode)
+
+    def mutate(self, info, id, input):
+        try:
+            group = Group.objects.get(pk=get_id(id))
+
+            if not group.can_change(info.context.user):
+                raise UserNotAuthorized
+
+            with reversion.create_revision():
+                user = User.objects.get(pk=get_id(input['userid']))
+                group.join(user, input['type'])
+                group.save()
+
+                reversion.set_user(info.context.user)
+                reversion.set_comment("changeMembershipGroup mutation")
+
+            ok = True
+        except Exception as e:
+            logger.error('Exception in ChangeMembershipGroup: {}.'.format(e))
+            group = None
+            ok = False
+
+
+        return ChangeMembershipGroup(group=group, ok=ok)
+
+class RemoveMembershipGroup(graphene.Mutation):
+
+    class Arguments:
+        id = graphene.ID(required=True)
+        userid = graphene.ID(required=True)
+
+    ok = graphene.Boolean()
+    group = graphene.Field(lambda: GroupNode)
+
+    def mutate(self, info, id, userid):
+        try:
+            group = Group.objects.get(pk=get_id(id))
+
+            if not group.can_change(info.context.user):
+                raise UserNotAuthorized()
+
+            with reversion.create_revision():
+                user = User.objects.get(pk=get_id(userid))
+                group.leave(user)
+                group.save()
+
+                reversion.set_user(info.context.user)
+                reversion.set_comment("removeMembershipGroup mutation")
+
+            ok = True
+        except Exception as e:
+            logger.error('Exception in RemoveMembershipGroup: {}.'.format(e))
+            group = None
+            ok = False
+
+        return RemoveMembershipGroup(group=group, ok=ok)
 
 class Mutation(graphene.ObjectType):
     create_comment = CreateComment.Field()
@@ -213,5 +337,7 @@ class Mutation(graphene.ObjectType):
     create_group = CreateGroup.Field()
     update_group = UpdateGroup.Field()
     delete_group = DeleteGroup.Field()
+    change_membership_group = ChangeMembershipGroup.Field()
+    remove_membership_group = RemoveMembershipGroup.Field()
     join_group = JoinGroup.Field()
     leave_group = LeaveGroup.Field()
